Increase precision of int64

+++ src/sequencer/sequencer.h
@@ -53,8 +53,8 @@ extern double scheduler_unlock[SAMPLES];
 class Client {
  public:
   virtual ~Client() {}
-  virtual void GetTxn(TxnProto** txn, int txn_id, int seed) = 0;
-  virtual void GetDetTxn(TxnProto** txn, int txn_id, int seed) = 0;
+  virtual void GetTxn(TxnProto** txn, int txn_id, int64 seed) = 0;
+  virtual void GetDetTxn(TxnProto** txn, int txn_id, int64 seed) = 0;
 };
 
 class Sequencer {
